refactor pipeline and fix bugs

+++ scripts/transE_clean.sh
@@ -9,7 +9,7 @@ set -e
 
 # CLI args
 WORK_DIR=$1
-LAST_ROUND=$WORK_DIR/last_round
+LAST_ROUND=$WORK_DIR/last_round/
 
 if [ ! -d "$WORK_DIR" ];then
   exit 0
@@ -19,6 +19,9 @@ if [ ! -d "$LAST_ROUND" ];then
   mkdir "$LAST_ROUND"
 fi
 
-mv $WORK_DIR/*.txt $LAST_ROUND
-mv $WORK_DIR/train/train2id.txt $LAST_ROUND
-mv $WORK_DIR/checkpoint/transe.ckp $LAST_ROUND
+[ -f "$WORK_DIR/valid_hrt.txt" ] && mv $WORK_DIR/*.txt $LAST_ROUND
+[ -f "$WORK_DIR/invalid_hrt.txt" ] && mv $WORK_DIR/*.txt $LAST_ROUND
+[ -f "$WORK_DIR/train/train2id.txt" ] && mv $WORK_DIR/train/train2id.txt $LAST_ROUND
+[ -f "$WORK_DIR/train/test2id.txt" ] && mv $WORK_DIR/train/test2id.txt $LAST_ROUND
+[ -f "$WORK_DIR/checkpoint/transe.ckpt" ] && mv $WORK_DIR/checkpoint/transe.ckpt $LAST_ROUND
+[ -f "$WORK_DIR/transE_raw_hrts.txt" ] && mv $WORK_DIR/transE_raw_hrts.txt $LAST_ROUND
